Refactor Jurisdiction Assignment's directory structure
