+++ src/configs/__mocks__/env.ts
@@ -175,3 +175,5 @@ export const ENABLE_DYNAMIC_MDA = true;
 export const ENABLED_JURISDICTION_METADATA_IDENTIFIER_OPTIONS = 'COVERAGE,POPULATION,RISK,TARGET'.split(
   ','
 );
+
+export const ASSIGNMENT_PAGE_SHOW_MAP = false;
