+++ src/containers/pages/JurisdictionAssignment/helpers/JurisdictionTable/index.css
